add:maste,optionの項目をユーザ内でユニークに変更

+++ app/Http/Controllers/MastesController.php
@@ -7,32 +7,42 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
-use Storage;
+use App\Services\MasteService;
 
 use App\Models\Maste;
 use App\Models\Designer;
 use App\Models\Printer;
 use App\Models\Country;
 use App\Models\Repository;
+
 use App\Http\Requests\MasteRequest;
 
 use Excel;
+use Storage;
 
 class MastesController extends Controller
 {
-    public function index()
+    protected $services;
+
+    public function __construct(MasteService $service)
     {
+        $this->service = $service;
+    }
+
+    public function index(Request $request)
+    {
+        //return \Auth::user()->name;
         //$mastes = Maste::all();
-        $mastes = Maste::orderBy('title', 'asc')->get();
+        $mastes = \Auth::user()->Maste()->orderBy('title', 'asc')->get();
         return view('mastes.index', compact('mastes'));
     }
 
     public function create()
     {
-        $designers = Designer::lists("name","id");
-        $printers = Printer::lists("name","id");
-        $countries = Country::lists("name","id");
-        $repositories = Repository::lists("name","id");
+        $designers = \Auth::user()->Designer->lists("name","id");
+        $printers = \Auth::user()->Printer->lists("name","id");
+        $countries = \Auth::user()->Country->lists("name","id");
+        $repositories = \Auth::user()->Repository->lists("name","id");
 
         return view('mastes.create', compact('designers', 'printers', 'countries', 'repositories'));
     }
@@ -40,26 +50,34 @@ class MastesController extends Controller
     public function store(MasteRequest $request)
     {
         $maste = Maste::create($request->all());
-        $file_name = "images/" . $maste->id . ".png";
-        Storage::put('images/text.png', file_get_contents($request->file('image')));
+        $maste->user_id = \Auth::user()->id;
+        $maste->save();
+        //$file_name = "images/" . $maste->id . ".png";
+        //Storage::put('images/text.png', file_get_contents($request->file('image')));
         return redirect('mastes');
     }
 
     public function show($id)
     {
         $maste = Maste::findOrFail($id);
-        //$image = Storage::get('images/text.png');
-        //var_dump($image);
+        if($maste->user_id != \Auth::user()->id) {
+            return redirect()->back();
+        }
         return view('mastes.show', compact('maste', 'image'));
     }
 
     public function edit($id)
     {
-        $designers = Designer::lists("name","id");
-        $printers = Printer::lists("name","id");
-        $countries = Country::lists("name","id");
-        $repositories = Repository::lists("name","id");
+
         $maste = Maste::findOrFail($id);
+        if($maste->user_id != \Auth::user()->id) {
+            return redirect()->back();
+        }
+
+        $designers = \Auth::user()->Designer->lists("name","id");
+        $printers = \Auth::user()->Printer->lists("name","id");
+        $countries = \Auth::user()->Country->lists("name","id");
+        $repositories = \Auth::user()->Repository->lists("name","id");
         return view('mastes.edit', compact('maste', 'designers', 'printers', 'countries', 'repositories'));
     }
 
@@ -73,6 +91,10 @@ class MastesController extends Controller
     public function destroy($id)
     {
         $maste = Maste::findOrFail($id);
+        if($maste->user_id != \Auth::user()->id) {
+            return redirect()->back();
+        }
+
         $maste->delete();
         \Session::flash('flash_message', 'マステを削除しました。');
         return redirect('mastes');
@@ -81,7 +103,7 @@ class MastesController extends Controller
     public function excel()
     {
         //
-        $maste = Maste::all();
+        $maste = \Auth::user()->Maste;
 
         Excel::create('plants', function($excel) use($maste) {
             $excel->sheet('Sheet 1', function($sheet) use($maste) {
